+++ api/models.py
@@ -2,6 +2,7 @@
 
 # pylint: disable=no-member, too-few-public-methods
 
+import uuid
 from django.contrib.gis.db import models
 from django.core.validators import RegexValidator
 
@@ -40,7 +41,11 @@ class ParkingLot(models.Model):
     """Model representing a parking lot."""
 
     auth_code = models.CharField(
-        max_length=50, auto_created=True, unique=True, null=False, blank=False
+        max_length=50,
+        default=uuid.uuid4,
+        unique=True,
+        null=False,
+        blank=False,
     )
     name = models.CharField(max_length=100, unique=True, null=False, blank=False)
     address = models.CharField(max_length=255, unique=True, null=False, blank=False)
@@ -82,7 +87,11 @@ class ParkingSpot(models.Model):
 
     parking_lot = models.ForeignKey(ParkingLot, on_delete=models.CASCADE)
     auth_code = models.CharField(
-        max_length=50, auto_created=True, unique=True, null=False, blank=False
+        max_length=50,
+        default=uuid.uuid4,
+        unique=True,
+        null=False,
+        blank=False,
     )
     event_list = models.ForeignKey(
         EventList, on_delete=models.CASCADE, null=True, blank=True